nax: address Task 2.3 review findings
- Widen NAXFrag32::load(device) and store(device) ld param from `short`
  to `int`. The threadgroup overloads keep `short ld` since they only
  ever receive ld=32 from a 32x32 scratch slice. The safe/rows variants
  already used `int ld`. Fixes a silent narrowing in gemm_epilogue's
  aligned C-load when ldc > 32767.

- Drop now-redundant `short(ld)` and `short(ldc)` casts in NAXTile's
  kPacking==1 branches and gemm_epilogue.

- Replace literal `1024` and `32 * 32` in the per-simdgroup scratch
  slice arithmetic with `NAXFrag_::kFragRows * kFragCols` so the size
  computation and slice offset cannot drift.

- Add a comment in gemm_epilogue's unaligned kPacking==1 branch
  explaining that negative per-frag row_lim/col_lim values are
  intentional for frags past the tile edge — NAXFrag32::load_safe's
  `r < row_lim` test zero-fills via signed comparison.

Probes 9/9 OK on M4 Pro.

+++ mlx/backend/metal/kernels/steel/gemm/nax.h
@@ -727,7 +727,7 @@ struct NAXFrag32 {
   METAL_FUNC static void load(
       thread dtype_frag_t<T>& dst,
       const device U* src,
-      const short ld) {
+      const int ld) {
     constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
         32, 32, 32,
         /*transpose_left=*/false,
@@ -754,7 +754,7 @@ struct NAXFrag32 {
   METAL_FUNC static void store(
       const thread dtype_frag_t<T>& src,
       device U* dst,
-      const short ld) {
+      const int ld) {
     constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
         32, 32, 32,
         /*transpose_left=*/false,
@@ -1117,7 +1117,7 @@ struct NAXTile {
           NAXFrag_t::load(
               frag_at<idx_row.value, idx_col.value>(),
               src + m_off * ld + n_off,
-              short(ld));
+              ld);
         });
       });
     } else {
@@ -1146,7 +1146,7 @@ struct NAXTile {
           NAXFrag_t::store(
               frag_at<idx_row.value, idx_col.value>(),
               dst + m_off * ld + n_off,
-              short(ld));
+              ld);
         });
       });
     } else {
