+++ tools/probe_nax_frag32.py
@@ -226,7 +226,7 @@ struct NAXFrag32 {
   METAL_FUNC static void load(
       thread dtype_frag_t<T>& dst,
       const device U* src,
-      const short ld) {
+      const int ld) {
     constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
         32, 32, 32,
         /*transpose_left=*/false,
@@ -253,7 +253,7 @@ struct NAXFrag32 {
   METAL_FUNC static void store(
       const thread dtype_frag_t<T>& src,
       device U* dst,
-      const short ld) {
+      const int ld) {
     constexpr auto desc = mpp::tensor_ops::matmul2d_descriptor(
         32, 32, 32,
         /*transpose_left=*/false,
